Close #329 - loggerf.future.instances.logFuture => loggerf.instances.future.logFuture
