+++ modules/logger-f-core/shared/src/main/scala-3/loggerf/instances/future.scala
@@ -1,4 +1,4 @@
-package loggerf.future
+package loggerf.instances
 
 import effectie.core.FxCtor
 import loggerf.core.Log
@@ -7,25 +7,23 @@ import loggerf.logger.CanLog
 import scala.concurrent.{ExecutionContext, Future}
 
 /** @author Kevin Lee
-  * @since 2022-02-08
+  * @since 2022-02-09
   */
-object instances {
-  @SuppressWarnings(Array("org.wartremover.warts.ImplicitParameter"))
-  implicit def logFuture(
-    implicit EF: FxCtor[Future],
+object future {
+
+  given logFuture(
+    using EF: FxCtor[Future],
     canLog: CanLog,
     EC: ExecutionContext,
   ): Log[Future] =
     new LogFuture(EF, canLog, EC)
 
-  @SuppressWarnings(Array("org.wartremover.warts.ImplicitParameter"))
   final class LogFuture(
     override val EF: FxCtor[Future],
     override val canLog: CanLog,
     val EC: ExecutionContext,
   ) extends Log[Future] {
-    override def flatMap0[A, B](fa: Future[A])(f: A => Future[B]): Future[B] =
-      fa.flatMap(f)(EC)
+    override def flatMap0[A, B](fa: Future[A])(f: A => Future[B]): Future[B] = fa.flatMap(f)(using EC)
   }
 
 }
